cleaned up edit recipes code
